+++ src/TimestampColumn.h
@@ -71,7 +71,7 @@ namespace mssql
 		   dt.day = ts.day;
 	   }
 
-	   static const int64_t NANOSECONDS_PER_MS = 1e6;                  // nanoseconds per millisecond
+	   static const int64_t NANOSECONDS_PER_MS = static_cast<int64_t>(1e6);                  // nanoseconds per millisecond
 
     private:
